+++ src/Filesystem.php
@@ -61,27 +61,27 @@ class Filesystem extends BaseFilesystem
      *  a stream resource
      * @param int $dirMode Mode for the directory, if it does not exist
      * @param bool $ignoreErrors With `true`, errors will be ignored
-     * @return bool
+     * @return string
      * @throws \Symfony\Component\Filesystem\Exception\IOException
      */
-    public function createFile(string $filename, $data = null, int $dirMode = 0777, bool $ignoreErrors = false): bool
+    public function createFile(string $filename, $data = null, int $dirMode = 0777, bool $ignoreErrors = false): string
     {
         try {
             $this->mkdir(dirname($filename), $dirMode);
             $this->dumpFile($filename, $data);
 
-            return true;
+            return $filename;
         } catch (IOException $e) {
             if (!$ignoreErrors) {
                 throw $e;
             }
 
-            return false;
+            return '';
         }
     }
 
     /**
-     * Creates a tenporary file. Alias for `tempnam()` and `file_put_contents()`.
+     * Creates a temporary file. Alias for `tempnam()` and `file_put_contents()`.
      *
      * You can pass a directory where to create the file. If `null`, the file
      *  will be created in `TMP`, if the constant is defined, otherwise in the
@@ -98,9 +98,8 @@ class Filesystem extends BaseFilesystem
     {
         $filename = tempnam($dir ?: (defined('TMP') ? TMP : sys_get_temp_dir()), $prefix) ?: '';
         Exceptionist::isTrue($filename, 'It is not possible to create a temporary file', RuntimeException::class);
-        $this->createFile($filename, $data);
 
-        return $filename;
+        return $this->createFile($filename, $data);
     }
 
     /**
@@ -109,7 +108,7 @@ class Filesystem extends BaseFilesystem
      * @param string|array|bool $exceptions Either an array of filename or folder
      *  names to exclude or boolean true to not grab dot files/folders
      * @param bool $ignoreErrors With `true`, errors will be ignored
-     * @return array<array> Array of nested directories and files in each directory
+     * @return array<array<string>> Array of nested directories and files in each directory
      * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
      */
     public function getDirTree(string $path, $exceptions = false, bool $ignoreErrors = false): array
@@ -136,9 +135,7 @@ class Filesystem extends BaseFilesystem
 
             $finder->files()->in($path);
             if ($exceptions) {
-                $exceptions = array_map(function ($exception) {
-                    return preg_quote($exception, '/');
-                }, $exceptions);
+                $exceptions = array_map(fn($exception): string => preg_quote($exception, '/'), $exceptions);
                 $finder->notName('/(' . implode('|', $exceptions) . ')/');
             }
             $files = objects_map(array_values(iterator_to_array($finder->sortByName())), 'getPathname');
@@ -349,8 +346,8 @@ class Filesystem extends BaseFilesystem
      * To remove the directory itself and all its contents, use the
      *  `rmdirRecursive()` method instead.
      * @param string $dirname The directory path
-     * @param array|bool|string $exceptions Either an array of files to exclude
-     *  or boolean true to not grab dot files
+     * @param array<string>|bool|string $exceptions Either an array of files to
+     *  exclude or boolean true to not grab dot files
      * @param bool $ignoreErrors With `true`, errors will be ignored
      * @return bool
      * @see \Tools\Filesystem::rmdirRecursive()